clean up and organise views
